# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insider Assessment App",
        "description": "API documentation for Insider Assessment App",
        "version": "1.0.0"
    },
    "paths": {
        "/api/recipients": {
            "get": {
                "summary": "Get all sent message recipients",
                "operationId": "da6c6ee90eed824b951f32a1032f6653",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for paginated results",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all sent message recipients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Et aliquid inventore illum sit ratione..."
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "5063584555"
                                                    },
                                                    "sent": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-28T15:49:52.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-28T15:55:48.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/recipients?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/recipients?page=1"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/recipients"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipients/{id}": {
            "get": {
                "summary": "Get a recipient by ID",
                "operationId": "92886ab95c8b2c4aa57b7f9842b01165",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the recipient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipient found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Et aliquid inventore illum sit ratione..."
                                        },
                                        "phone_number": {
                                            "type": "string",
                                            "example": "5063584555"
                                        },
                                        "sent": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-28T15:49:52.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-28T15:55:48.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipient not found"
                    }
                }
            }
        }
    }
}